# feat(values): reduce memory allocations for operations in values