@jsternberg 2019-12-09
Benchmark comparison:

```
jsternberg:[~/g/s/g/i/flux] $ (feat/optimize-object-set $) benchcmp old.txt new.txt
benchmark                          old ns/op       new ns/op       delta
BenchmarkFilter_Values/1000-12     70260034867     21914358279     -68.81%

benchmark                          old allocs     new allocs     delta
BenchmarkFilter_Values/1000-12     905505262      155743661      -82.80%

benchmark                          old bytes       new bytes      delta
BenchmarkFilter_Values/1000-12     21482642584     6985710944     -67.48%
```

There's still work to do here, but this will have to be good enough for now until I've got some more cycles for dealing with performance related stuff.